[PIL-2221] - change class name to reflect repayment
