+++ app/controllers/repayments/ReasonForRequestingRepaymentController.scala
@@ -18,7 +18,7 @@ package controllers.repayments
 
 import config.FrontendAppConfig
 import controllers.actions._
-import forms.ReasonForRequestingRefundFormProvider
+import forms.ReasonForRequestingRepaymentFormProvider
 import models.Mode
 import navigation.RepaymentNavigator
 import pages.ReasonForRequestingRefundPage
@@ -34,14 +34,14 @@ import javax.inject.{Inject, Named}
 import scala.concurrent.{ExecutionContext, Future}
 
 class ReasonForRequestingRepaymentController @Inject() (
-  val sessionRepository:                  SessionRepository,
-  @Named("EnrolmentIdentifier") identify: IdentifierAction,
-  getData:                                SessionDataRetrievalAction,
-  navigator:                              RepaymentNavigator,
-  requireData:                            SessionDataRequiredAction,
-  formProvider:                           ReasonForRequestingRefundFormProvider,
-  val controllerComponents:               MessagesControllerComponents,
-  view:                                   ReasonForRequestingRefundView
+                                                         val sessionRepository:                  SessionRepository,
+                                                         @Named("EnrolmentIdentifier") identify: IdentifierAction,
+                                                         getData:                                SessionDataRetrievalAction,
+                                                         navigator:                              RepaymentNavigator,
+                                                         requireData:                            SessionDataRequiredAction,
+                                                         formProvider:                           ReasonForRequestingRepaymentFormProvider,
+                                                         val controllerComponents:               MessagesControllerComponents,
+                                                         view:                                   ReasonForRequestingRefundView
 )(implicit ec:                            ExecutionContext, appConfig: FrontendAppConfig)
     extends FrontendBaseController
     with I18nSupport {
